Add mangohud-next alongside mangohud

+++ build.sh
@@ -151,8 +151,9 @@ configure() {
     if [[ ! -f "build/meson32/build.ninja" && "$MACHINE" = "x86_64" ]]; then
         export CC="gcc -m32"
         export CXX="g++ -m32"
+        export LDFLAGS=-m32
         export PKG_CONFIG_PATH="/usr/lib32/pkgconfig:/usr/lib/i386-linux-gnu/pkgconfig:/usr/lib/pkgconfig:${PKG_CONFIG_PATH_32}"
-        meson setup build/meson32 --libdir lib/mangohud/lib32 --prefix /usr -Dappend_libdir_mangohud=false $@ ${CONFIGURE_OPTS}
+        meson setup build/meson32 --libdir lib/mangohud/lib32 --prefix /usr -Dappend_libdir_mangohud=false -Dwith_server=false $@ ${CONFIGURE_OPTS}
     fi
 }
 
@@ -198,7 +199,11 @@ uninstall() {
     rm -fv "/usr/share/vulkan/implicit_layer.d/MangoHud.json"
     rm -fv "/usr/share/vulkan/implicit_layer.d/MangoHud.x86.json"
     rm -fv "/usr/share/vulkan/implicit_layer.d/MangoHud.x86_64.json"
+    rm -fv "/usr/share/vulkan/implicit_layer.d/MangoHud-next.x86.json"
+    rm -fv "/usr/share/vulkan/implicit_layer.d/MangoHud-next.x86_64.json"
     rm -fv "/usr/bin/mangohud"
+    rm -fv "/usr/bin/mangohud-next"
+    rm -fv "/usr/bin/mangohud-server"
     rm -fv "/usr/bin/mangoplot"
     rm -fv "/usr/bin/mangohud.x86"
 }
@@ -231,17 +236,35 @@ install() {
     /usr/bin/install -Dvm644 ./build/release/usr/lib/mangohud/lib64/libMangoHud.so /usr/lib/mangohud/lib64/libMangoHud.so
     /usr/bin/install -Dvm644 ./build/release/usr/lib/mangohud/lib64/libMangoHud_opengl.so /usr/lib/mangohud/lib64/libMangoHud_opengl.so
     /usr/bin/install -Dvm644 ./build/release/usr/lib/mangohud/lib64/libMangoHud_shim.so /usr/lib/mangohud/lib64/libMangoHud_shim.so
+    if [[ -f ./build/release/usr/lib/mangohud/lib64/libMangoHud-next.so ]]; then
+      /usr/bin/install -Dvm644 ./build/release/usr/lib/mangohud/lib64/libMangoHud-next.so /usr/lib/mangohud/lib64/libMangoHud-next.so
+    fi
     if [ "$MACHINE" = "x86_64" ]; then
       /usr/bin/install -Dvm644 ./build/release/usr/lib/mangohud/lib32/libMangoHud.so /usr/lib/mangohud/lib32/libMangoHud.so
       /usr/bin/install -Dvm644 ./build/release/usr/lib/mangohud/lib32/libMangoHud_opengl.so /usr/lib/mangohud/lib32/libMangoHud_opengl.so
       /usr/bin/install -Dvm644 ./build/release/usr/lib/mangohud/lib32/libMangoHud_shim.so /usr/lib/mangohud/lib32/libMangoHud_shim.so
+      if [[ -f ./build/release/usr/lib/mangohud/lib32/libMangoHud-next.so ]]; then
+        /usr/bin/install -Dvm644 ./build/release/usr/lib/mangohud/lib32/libMangoHud-next.so /usr/lib/mangohud/lib32/libMangoHud-next.so
+      fi
     fi
 
     /usr/bin/install -Dvm644 ./build/release/usr/share/vulkan/implicit_layer.d/MangoHud.x86_64.json /usr/share/vulkan/implicit_layer.d/MangoHud.x86_64.json
     /usr/bin/install -Dvm644 ./build/release/usr/share/vulkan/implicit_layer.d/MangoHud.x86.json /usr/share/vulkan/implicit_layer.d/MangoHud.x86.json
+    if [[ -f ./build/release/usr/share/vulkan/implicit_layer.d/MangoHud-next.x86_64.json ]]; then
+      /usr/bin/install -Dvm644 ./build/release/usr/share/vulkan/implicit_layer.d/MangoHud-next.x86_64.json /usr/share/vulkan/implicit_layer.d/MangoHud-next.x86_64.json
+    fi
+    if [[ -f ./build/release/usr/share/vulkan/implicit_layer.d/MangoHud-next.x86.json ]]; then
+      /usr/bin/install -Dvm644 ./build/release/usr/share/vulkan/implicit_layer.d/MangoHud-next.x86.json /usr/share/vulkan/implicit_layer.d/MangoHud-next.x86.json
+    fi
     /usr/bin/install -Dvm644 ./build/release/usr/share/man/man1/mangohud.1 /usr/share/man/man1/mangohud.1
     /usr/bin/install -Dvm644 ./build/release/usr/share/doc/mangohud/MangoHud.conf.example /usr/share/doc/mangohud/MangoHud.conf.example
     /usr/bin/install -vm755  ./build/release/usr/bin/mangohud /usr/bin/mangohud
+    if [[ -f ./build/release/usr/bin/mangohud-next ]]; then
+      /usr/bin/install -vm755  ./build/release/usr/bin/mangohud-next /usr/bin/mangohud-next
+    fi
+    if [[ -f ./build/release/usr/bin/mangohud-server ]]; then
+      /usr/bin/install -vm755  ./build/release/usr/bin/mangohud-server /usr/bin/mangohud-server
+    fi
     /usr/bin/install -vm755  ./build/release/usr/bin/mangoplot /usr/bin/mangoplot
 
     ln -sv $DEFAULTLIB /usr/lib/mangohud/lib
@@ -290,6 +313,7 @@ clean() {
     rm -rf "build"
     for path in subprojects/*/; do
         [[ $path == "subprojects/packagefiles/" ]] && continue
+        [[ $path == "subprojects/packagecache/" ]] && continue
         rm -rf "$path"
     done
 }
